mv get*EnergyDepositionConstants to blocks.py
- in ARMI and at least one major downstream project, these are only every used on blocks. So these can be safely moved off of composites.py and onto blocks.py.
- also, given the current properties, only Blocks are even able to use these methods. assembly.r and component.r don't exist so they wouldn't work anyway. and Core.getMicroSuffix doesn't exist so Core wouldn't work either. These methods really only work for blocks, so let's move them to blocks.py.
